refactor: 북마크 save·merge 모달을 bookmark-modals.js로 분리 (ADR-034 후속 PR5c)
모달 분리 3단계. save/edit 모달(openSaveModal·_showSaveModal·commitSaveBookmark)
과 merge 다이얼로그(상호 결합 — save→merge, merge→_showSaveModal)를 함께
bookmark-modals.js로 이동. 콤보박스(PR5b)가 이미 모듈에 있어 save 의 폴더 picker
호출이 모듈 내부 호출이 됐다.

DI 에 exitVerseSelectMode 추가(절 저장·합치기 후 선택 모드 종료). Escape 스택에
merge·save 를 closeTopmostModal 로 흡수 — move/import 만 bookmark.js 잔류(PR5d).
모달은 new-folder 외 상호 배타라 경계 너머 순서 무관(주석 명시).

이동으로 미사용이 된 import 정리: bookmark.js 의 collapseSegmentedVerses·
mergeVerseSpecs(verse-spec), closeConfirmModal·closeChapterDeleteModal(modals).
modals.js export 를 bookmark.js 가 실제 호출하는 진입점으로 축소.

bookmark.js 2659 -> 2393 줄, bookmark-modals.js 498 -> 778 줄. 누적 3578 -> 2393 (-33%).
